commited

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -1,5 +1,4 @@
 #include "lists.h"
-#include <stdint.h>
 #include <stdio.h>
 
 /**
@@ -11,18 +10,18 @@
 size_t print_listint_safe(const listint_t *head)
 {
     size_t num = 0;
-    intptr_t diff;
+    long int diff;
 
     while (head)
     {
-        diff = (intptr_t)head - (intptr_t)head->next;
+        diff = head - head->next;
         num++;
-        printf("[%p] %d\n", (void *) (intptr_t) head, head->n);
+        printf("[%p] %d\n", (void *)head, head->n);
         if (diff > 0)
             head = head->next;
         else
         {
-            printf("-> [%p] %d\n", (void *) (intptr_t) head->next, head->next->n);
+            printf("-> [%p] %d\n", (void *)head->next, head->next->n);
             break;
         }
     }
